pyproject.toml
[project]
name = "framework"
version = "0.1.0"
description = "Multi-agent orchestration framework"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "anthropic>=0.40",
    "rich>=13.7",
    "pyyaml>=6.0",
    "pydantic>=2.6",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
framework = "framework.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["framework*"]

[tool.setuptools.package-data]
framework = ["schema.sql", "templates/*.md", "templates/agents/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
